02_can_data_preprocessing/requirements.txt
cantools==39.4.5
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pyaml==6.0.1
pyarrow==16.1.0
scipy==1.13.1
tqdm==4.66.4